Moved TestApi code into its own folder
The stuff for API helpers living with the other helpers didn't make a
ton of sense so moving it out into its own place.
