+++ LeaderboardBackend.Test/TestApi/TestApiClient.cs
@@ -5,8 +5,9 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text.Json;
 using System.Threading.Tasks;
+using LeaderboardBackend.Test.Lib;
 
-namespace LeaderboardBackend.Test.Lib;
+namespace LeaderboardBackend.Test.TestApi;
 
 internal sealed class RequestFailureException : Exception
 {
