+++ LeaderboardBackend.Test/Users.cs
@@ -4,11 +4,12 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using LeaderboardBackend.Models.Requests;
 using LeaderboardBackend.Models.Entities;
+using LeaderboardBackend.Test.TestApi;
+using LeaderboardBackend.Test.TestApi.Extensions;
 using System.Net.Http.Json;
 using System;
 using System.Net.Http.Headers;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using LeaderboardBackend.Test.Lib;
 
 namespace LeaderboardBackend.Test;
